fix(v0.31.0): address code review MAJORs — common test module + LLM contract pins + CHANGELOG
Closes w4-code-review findings:
- MAJOR-1: extract McpFixture к crates/forgeplan-mcp/tests/common/mod.rs
  (was duplicated verbatim between integration_e2e.rs и integration_full_coverage.rs)
- MAJOR-2: tighten LLM-dependent MCP tests с is_error=true assertion
  (capture/reason/decompose/generate — guaranteed error без LLM provider)
  Remaining 30+ tolerant tests deferred к follow-up (PROB-065 candidate)
- MAJOR-3: populate CHANGELOG.md [Unreleased] с PROB-064 + openssl + tests entries

Bonus: delete dead watch_smoke test (MINOR-1 — empty #[ignore] body, no value)

Pipeline gates все green: fmt/check/clippy/test/smoke

Refs: w4-code-review findings, PROB-064, PROB-065 (TBD)
